requirements.txt
cachetools==4.1.1
certifi==2020.6.20
cffi==1.14.1
chardet==3.0.4
colorama==0.4.3
gevent==20.6.2
google-api-core==1.22.0
google-api-python-client==1.10.0
google-auth==1.20.0
google-auth-httplib2==0.0.4
google-auth-oauthlib==0.4.1
googleapis-common-protos==1.52.0
greenlet==0.4.16
httplib2==0.18.1
idna==2.10
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
memory-profiler==0.57.0
oauth2client==4.1.3
oauthlib==3.1.0
protobuf==3.12.4
psutil==5.7.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
pycparser==2.20
Pympler==0.8
pyprof2calltree==1.4.5
pytz==2020.1
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.6
signalr-client==0.0.7
six==1.15.0
sseclient==0.0.26
uritemplate==3.0.1
urllib3==1.25.10
websocket-client==0.57.0
websockets==8.1
wincertstore==0.2
wrapt==1.11.2
zope.event==4.4
zope.interface==5.1.0
